feat(secret-rotation): changed to mysql2 client and refactored queue util functions
